pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "google-genai>=1.56.0",
    "llama-index-core>=0.14.10",
    "llama-index-llms-openai>=0.3.0",
    "llama-index-embeddings-openai>=0.3.0",
    "python-dotenv>=1.0.0",
    "pypdf>=5.0.0",
    "streamlit>=1.52.2",
    "llama-index-llms-openai-like>=0.5.3",
    "llama-index-readers-file>=0.5.5",
    "pymupdf>=1.26.7",
    "ultralytics>=8.3.241",
    "opencv-python-headless>=4.12.0.88",
]
